xml extravaganza
- made it so there's a 200ms leniency between hitting jump and landing
(bunnyhopping is easier, basically, but still not like quake)
- jump now goes to next level
- added markup detection to signs?! holy wow this is a big commit. so
you can $do this$ to text to make it orange ok
- disabled input resetting on state change so you can hold left/right to
instantly move upon new state load
- TILED MAP now supports "collection of images" tilemaps! it currently
supports one, that must be named "props". any "tiles" added to this can
be played via "tile objects" in the editor and their
scale/rotation/flips are all respected in the game. this is pretty much
the replacement for tilesdetail
- removed camera scrollbounds
- oh yeah, levels are smaller and stuff now
- just have a grey background until i figure out what to do with
parallax background ... i don't want to have a bunch of tilemap padding
for no reason, so maybe in the future i'll manually make flxsprite
padding