props, cleanup some mess, make more mess
- found out that props actually weren't flipping correctly, but also
found the reason why after much trial and tribulation, it's because
o.gid is just an Int and i need it to be unsigned or 64 bit. not sure
what im going to do here, it works on flash anyway and i found a hacky
solution for windows. neko is broken pending i update TiledObject.hx.
- made props have a scrollx property in tiled which changes their
scrollfactor.x. the props are sorted automatically by this value and
added to various groups so to be properly sorted.
- rockets are now z-sorted behind the player and muzzleflash. subtle but
big improvement!
- bit the bullet and added big dumb solid color flxsprite map borders
now that we have parallax objects. Flash Hates This (re:
https://groups.google.com/forum/#!topic/haxeflixel/xeGbzgoRLJ0)