found cause of innerHitbox drama, refactored death
so innerhitbox being never added to state means it wasn't updating. THIS
is why adding it worked for one of the bugs. But then it didn't fix the
other bug, which was caused by the updating being out of sync with the
player. so, the solution is to NOT add it to state, and call update()
manually on it alongside player's super update call. Phew.

oh and player death function is no longer awful